refactor(analytics): separate controller from content
